+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -254,6 +254,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-slot-keylet-buf-len
+  <clang-tidy/checks/hooks-slot-keylet-buf-len>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-slot-limit
   <clang-tidy/checks/hooks-slot-limit>` check.